requirements.txt
altgraph==0.17
asgiref==3.3.1
auto-py-to-exe==2.7.11
bottle==0.12.19
bottle-websocket==0.2.9
dj-database-url==0.5.0
Django==3.1.5
django-filter==2.4.0
django-rest-multiple-models==2.1.3
djangorestframework==3.12.2
Eel==0.12.4
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
gunicorn==20.0.4
Markdown==3.3.3
numpy==1.19.5
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pypi==2.1
pytz==2020.5
sqlparse==0.4.1
whichcraft==0.6.1
whitenoise==5.2.0
zope.event==4.5.0
zope.interface==5.2.0
